Changed a lot of small things and some bug fixes. changed card numbers to long, fixed a lot of typos in the output messages, added stuff to the change base rate function and email function in ophelias oasis, used the class records variable at the top of the class for all my calls to records functions, added a change res type field in edit reservation and added check in functionality in the check in function in reservations class. There are other small changes here and there as well. feel free to contact me with any questions.

+++ OpheliasOasis/src/opheliasoasis/CreditCard.java
@@ -17,10 +17,10 @@ public class CreditCard implements Serializable {
     String CardHolder;
     int expMonth;
     int expYear;
-    int cardNumber;
+    long cardNumber;
     int CSV;
     
-    public CreditCard(String CardHolder, int expMonth, int expYear, int cardNumber, int CSV){
+    public CreditCard(String CardHolder, int expMonth, int expYear, long cardNumber, int CSV){
         this.CardHolder = CardHolder;
         this.expMonth = expMonth;
         this.expYear = expYear;
@@ -41,7 +41,7 @@ public class CreditCard implements Serializable {
         return expYear;
     }
     
-    public int getCardNumber(){
+    public long getCardNumber(){
         return cardNumber;
     }
     
@@ -57,7 +57,7 @@ public class CreditCard implements Serializable {
     public void setExpYear(int NewYear){
         expYear = NewYear;
     }
-    public void setCardNumber(int NewCardNumber){
+    public void setCardNumber(long NewCardNumber){
         cardNumber = NewCardNumber;
     }
     public void setCSV(int NewCSV){
